requirements.txt
anthropic>=0.49
python-telegram-bot==21.5
supabase>=2.28
apscheduler==3.10.4
pyyaml>=6.0
python-dotenv>=1.0
httpx>=0.27,<0.28
pytest
pytest-asyncio
